Blocking facebook by address list
There are many ways to block facebook requests. This script can be used to block facebook using an address list. Be aware that you need a firewall rule in order to block requests based on the address list. Other options are setup a dstnat rule to forward all the DNS requests either TCP or UDP at destination port 53 and forward it to the router itself, from there it is possible to add static entries.